pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "remora-vibecheck"
version = "0.1.0"
description = "Scaffold for ragged batching and W8A16 quantization experiments"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Remora Team" }]
dependencies = [
    "torch>=2.2",
    "transformers>=4.38",
    "pillow>=9.0",
]

[project.optional-dependencies]
eval = ["vlmeval"]

[tool.setuptools.packages.find]
include = ["remora"]

[tool.setuptools]
py-modules = ["vibecheck"]
